+++ Cargo.toml
@@ -41,7 +41,6 @@ tokio = { version = "1.42.0", default-features = false, features = [
 tokio-util = { version = "0.7.13", default-features = false, features = ["rt"] }
 tracing = { version = "0.1.41", default-features = false }
 tracing-subscriber = { version = "0.3.19", features = ["env-filter"] }
-chrono = "0.4"
 
 [profile.test.package.proptest]
 opt-level = 3
