+++ crates/p2p/Cargo.toml
@@ -31,5 +31,4 @@ tokio = { workspace = true, features = ["rt", "rt-multi-thread"] }
 tokio-util.workspace = true
 tracing-subscriber.workspace = true
 anyhow = "1.0.95"
-chrono = "0.4"
 rand = "0.9.1"
