+++ crates/p2p/src/tests/mod.rs
@@ -3,4 +3,4 @@ pub(crate) mod gossipsub;
 pub(crate) mod is_connected;
 pub(crate) mod new_op;
 
-pub(crate) mod three_handlers;
+pub(crate) mod request_response;
